Support OpenBSD (#6)
* remove hard-coded strings

* support OpenBSD and Ubuntu

* support CentOS

* wait for es to start

* create dependency only when platform is RedHat

* new box has been released

* skip OpenBSD

new box with X installed is being built now

* restore OpenBSD

* wait_for the service instead of calling a handler

the order of handler executions is FIFO and multiple calls to a hanlder
are aggregated to one handler execution. the initial intension of the
notify is to wait immediately after the "Start" task. instead of calling
the handler, "when" and "changed" ensure that the play always wait right
after the "Start" task.